+++ projects/design-angular-kit/schematics/ng-add/setup-project/rules/index.ts
@@ -1,3 +1,4 @@
 export * from './add-animations';
 export * from './add-design-angular-kit';
 export * from './add-http-client';
+export * from './add-import-to-style-file';
